[PATCH] 3c59x suspend/resume fix
Currently, all of the 3c59x power management code is disabled unless the
`enable_wol' module parameter is provided.  This was done because the PM
support was added quite late in the 2.4 cycle.

It was always intended that this conditionality be removed in 2.5.

BKrev: 3f2a9b88-FaPaCQQXe8DnKpi8sw3vQ