[PATCH] AS barrier bug
There's a bug in AS, it sets HARDBARRIER on a request when it should be
setting it as a softbarrier. IO scheduler has no business using that
bit, if it needs to put barriers there they should be soft. Hard
barriers involve hardware operations.

BKrev: 3f2e8a81tf4LNyZSwy__m62Zch27Xw