[PATCH] Make sure sunrpc/cache doesn't confuse writers with readers.
When a sunrpc/cache channel is not open for reading, the
cache doesn't bother making and waiting for up-calls.

However it doesn't currently distingish between open-for-read/write
and open-for-write, so an op-for-write will look like a reader
and will cause inappropriate waiting.

This patch checks if a file is open-for-read and will only register
a file as a reader if it really is one.

BKrev: 3f301148ZBtzZOmYzXbPDjSc8PrEnQ