Merge bk://bk.arm.linux.org.uk/linux-2.5-rmk into home.osdl.org:/home/torvalds/v2.5/linux
2003/08/05 22:02:32+01:00 rmk
[ARM] Add asm-arm/local.h

2003/08/05 21:44:58+01:00 rmk
[ARM] Fix platform_add_devices() prototype

2003/08/05 21:22:57+01:00 rmk
[ARM] Fix two Acorn SCSI drivers

ARXESCSI does not have the interrupt routed to the host
machine, so we need to babysit the host interface.  We
do this via a variant of the normal command function in
fas216.c

Fix a bug in the procfs file handling in the EESOX SCSI
driver.

2003/08/05 20:39:43+01:00 rmk
[ARM] Remove more static register accesses from sa1111.c

Use the ioremapped region for accesses to SA1111 IRQ
controller registers.

2003/08/05 20:27:31+01:00 rmk
[ARM] Update machine types file.

2003/08/05 20:20:43+01:00 rmk
[ARM] Fix SSP IRQ handler return type.

2003/08/05 20:16:48+01:00 rmk
[ARM] Remove unused get_hackkit_scr prototype

BKrev: 3f301d90op16H08MQ9Q0f51nSrECdQ