Merge bk://linux-pnp.bkbits.net/pnp-2.5 into home.osdl.org:/home/torvalds/v2.5/linux
2003/08/05 21:22:55+00:00 ambx1
[PNP] Increment Version Number

2003/08/05 21:21:39+00:00 ambx1
[PNP] Remove device naming based on id

This patch removes the pnp name database code.  Most buses, including
pnp, will be using userspace to name devices in the near future.  Also
dev->name will be removed from the driver model soon.

2003/08/05 20:34:05+00:00 ambx1
[PNPBIOS] Move low level code into a separate file

This patch moves the low level bios calls to a separate file,
"bioscalls.c".  It is a cleanup that will improve organization of the
pnpbios driver code.

2003/08/05 20:18:25+00:00 ambx1
[SOUND] Remove __(dev)initdata from all pnp sound drivers

This patch is needed in order to avoid a potential oops.  It is
similiar to the changes made to pci.


2003/08/05 20:15:00+00:00 ambx1
[PNPBIOS] Move Parsing Functions to the PnPBIOS driver

This patch moves the resource parsing functions from support.c to the
pnpbios driver.  Originally these functions were intended for other
pnp protocols but in reality they are only used by the PnPBIOS driver.
This patch greatly cleans up the code in both the parsing functions
and their connection with the pnpbios driver.  Also note that
pnpbios.h has been added for local pnpbios functions.

2003/08/05 14:36:01+00:00 ambx1
[PNP] Remove protocol_data from pnp_dev and pnp_card

This is not needed.


2003/08/05 14:29:11+00:00 ambx1
  [PNP] Handle unset resources properly

  This patch is similar to the disabled resource patch in that it
  avoids direct numeric comparisons with data in unset resource
  structures.

BKrev: 3f3086a3Lwnf88ARsyW2Gkvq-_9udw