[PATCH] initialise page->private
From: Nathan Scott <nathans@sgi.com>

XFS wants to use page->private as a bitmap of uptodate indicators for
sub-page-sized blocks (which is one of the things ->provate was intended
for).

But someone needs to initialise ->private somewhere.  best to do it in the
page allocator, so the zeroness of a new page's ->private becomes a
system-wide thing.

BKrev: 3f329734YYZ9lqi16j2ls1KTeHSLJQ