[PATCH] fix trident.c lockup on module load 2.6.0-test2
This patch fixes a kernel lockup with 2.6.0-test2 when the trident.c
OSS driver is loaded and the driver attempts to initialize the
card. The problem is that in ali_ac97_get() we lock the card->lock
spinlock, but never release it on the good path, only on the error
path. This patch adds the missing spin_unlock_irqrestore().

This bug snuck in in a 2.4 sync from Alan, and 2.4 appears to suffer
from the same problem.  A patch for that will be send to Marcelo
momentarily.

BKrev: 3f3b2e42DV4IxHdsKitTSAMEsgr9VA