[IRDA]: Get rid of useless hashbin in irtty.
Working on converting IRDA to new netdevice semantics, this is the first
of some of the small things I found.

The irtty hashbin is created, maintained and never used.
Motivation for removing is that the hashbin locking has problems with
the locking assumptions in on network device removal.

BKrev: 3f3e48a2NR1eT-TcmRp-VpxD3F1DSA