[PATCH] Move config tasks to kconfig/Makefile
This fixes a bug with multiple targets.

Olaf Hering reported that the build failed for PowerPc if used like
this: make oldconfig zImage

The reason for this was that .config was not present for any targets
specified in arch/$(ARCH)/Makefile and below.

That's because .config would not be included when oldconfig is present
in the list of targets.  The fix is to move handling of *config task to
the kconfig/Makefile.  Furthermore the logic in top-level makefile has
changed a bit, creating a more logial structure.  When building a fresh
kernel, the user is now told that .config is missing, not an anonymous
report that .config did not exist.

The error has survided this long because the targets used in i386/boot
in general does not use CONFIG_ symbols.

Olaf Hering has tested this patch with success.

BKrev: 3f400fc4xrKe4htXvP_B-Ul2t8EB_w