[PATCH] add ASUS l3800P to DMI black list
From: Eric Valette <eric.valette@free.fr>

The following patch integrated in 2.5.74,

<http://lists.insecure.org/lists/linux-kernel/2003/Jun/5840.html>

really enables the APIC even if BIOS disabled it. Unfortunately,
enabling APIC really does not seem to work on this ASUS laptop and ACPI
(which is mandatory) crash the kernel in ACPI code at boot time while
"Executing all Devices _STA and_INIT methods"

Unless someones find a bug in ACPI code related to APIC management, It
is safer to add this machine in the DMI black list (along with DELL,
IBM, ...).

So, as suggested by the author of the problematic change, I added and
entry in the DMI black list. But my guess is that most laptop will soon
be present in this list....

BKrev: 3f42430cdIjYhNk_PS6R1eXLsbffQQ