[PATCH] Fix awe PnP probing
This is a rewrite of the awe_wave detection code that will allow this
driver to be compiled.  It moves detection functions to a common
location at the end of the file and makes the code driver-model
compatible.  Also it fixes a bug in which the driver could possibly
write to incorrect ports when using isapnp cards.

Unfortunantly I do not currently have an AWE32 to test these changes so
I could only check for compilation and driver registration.

BKrev: 3f425b06qjUp_FaK6WcrIkmJFFF8Ew