[PATCH] correct local_dec on some architectures
A small bug in local.h apparently got copied a few times.
I noticed this because I copied the same bug to s390.
This patch should fix the occurrences in BK, but there
are others that are not merged yet, e.g. ppc64 in -mm3.

BKrev: 3f425f7fbGwohZAyZZWqPQsJ2z8ezQ