[PATCH] fix intel copy_to_user()
A `cat /dev/kmem' oopses the kernel.  This is because new Intel
copy_to_user() doesn't handle segfaults against the kernel-side source
address of the copy.

It is not obvious how to fix read_kmem(), and handling faults on either
source or dest is traditional behaviour, so fix it in the copy function by
adding exception table entries for the reads as well as the writes.

BKrev: 3f447493QFba3EzguMU-Q-YouM542g