[PATCH] fix for htree corruption
From: chrisl@vmware.com

Fixes the "perl installer weirdness" problem which a few people have been
reporting.

The bug is triggered by creating the index.  Coping out the index we assume
the dirents start with the first entry after "." "..".

It can make the first previous deleted entry reappear.  In the past we set
inode to zero for an empty entry so this is not a problem.  That is not
true any more.

Also, there were a few places where we were altering inode times after
running ext3_mark_inode_dirty().  We should be doing that beforehand.

BKrev: 3f4474acydhN7spsoEfsO2X_l2OHxA