[PATCH] disable prefetch on athlons
K7's (at least) are faulting in the prefetch instruction.  The AMD
engineers have said they will be getting back to us on it, and the fix is
looking complex, and nobody seems to be standing up to work on it.

So hum.  The usual manifestation is an oops in hlist_for_each(), down in
the VFS inode lookup code.  Disrupting our testers in this way is very bad,
so this patch just disables prefetch on all AMD parts in a rather stupid
way.

BKrev: 3f52a02cU_TyRcA8sMaNQzdt6GZrMg