[PATCH] Use tgid rather than pid in dnotify
From: Ulrich Drepper <drepper@redhat.com>

I'm not entirely sure about this change.  But it seems to be necessary.
The dnotify code stores the PID in the file structure.  The entire process
shares the file and any signal (is it used for that?) should be sent to the
process (thread group), not the individual thread.  Also keep in mind that
threads can go away while the process (and therefore file descriptor)
remain.  And the ID of the thread can be reused.

Somebody who knows this code should take a good look.

(Looks right to me...)

BKrev: 3f52a213Jzam3Nu8DjV6uXGpejwetQ