Merge bk://kernel.bkbits.net/jgarzik/net-drivers-2.5 into home.osdl.org:/home/torvalds/v2.5/linux
2003/09/03 13:53:05-04:00 javier
[wireless airo] build fix when MIC support is disabled

2003/09/03 13:43:41-04:00 ernstp
[netdrvr] list CONFIG_BMAC in drivers/net/Makefile.lib,
as it uses the crc32 library.

2003/09/03 13:39:27-04:00 mzyngier
[netdrvr de4x5] big modernization / cleanup

The enclosed patch tries to clean the tulip de4x5 driver, and update
it to some of the 2.6 APIs :

- Use PCI and EISA probing APIs
- Use generic DMA API
- Fix DE425 init on the Jensen
- Remove de4x5 from Space.c

It's been tested on x86 and Alpha, with a DE425 (21040, EISA), a DE435
(21040, PCI) and a quad DLink (4*21143, PCI).

The major problem with this patch is that, because of the Space.c
removal, interfaces get potentially renumbered. This has been
discussed to death on lkml, without any obvious solution. IMHO, we'd
better remove as many drivers as we can from Space.c before
2.6.0-final ships...

2003/09/03 12:35:56-04:00 bunk
[wireless airo] fix build with gcc 2.95

BKrev: 3f562eb9vJ6V8fmpuGrCJ1XXTcljUQ