[PATCH] i8042 free_irq() aliasing fix
The same address `i8042_request_irq_cookie' is used in three places for the
i8042 request_irq() argument.  This means that if someone calls
i8042_check_mux() or i8042_check_aux() while the IRQ is in use, the
free_irq() call in there will free the wrong IRQ handler.

So give all three instances of request_irq() in i8042.c a distinct address by
which to identify the IRQ instance.

(This is probably a non-bug, because the `check' functions are not called
when the device is open, but it is better this way).

BKrev: 3f56f433XP7_DANrwxp2DKENeeaNUQ