Merge bk://linux-dj.bkbits.net/cpufreq into home.osdl.org:/home/torvalds/v2.5/linux
2003/09/05 15:38:20+01:00 davej
[CPUFREQ] Print out CPU name in debug info.
Add template for Nehemiah.

2003/09/05 15:19:00+01:00 davej
[CPUFREQ] Print out FSB in longhaul debug info

2003/09/04 03:23:06+01:00 davej
[CPUFREQ] Move longhaul scale/ratio tables to longhaul header file

2003/09/04 03:02:49+01:00 davej
[CPUFREQ] Remove the voltage scaling from longhaul driver.
The 'match voltage to frequency' routine is complete crap and needs
rewriting. (It also only works for a single stepping of C3)
I'll redo it when I find the specs again.
In the meantime, it doesn't work properly, so kill it.

2003/09/04 02:51:33+01:00 davej
[CPUFREQ] Inform user of status of Ezra-T/Nehemiah longhaul support.
These are currently unsupported, and I get quite a lot of mails
asking about it. Hopefully this will be a little more informative
than 'unknown CPU'.

2003/09/04 02:38:00+01:00 davej
[CPUFREQ] Clean up clock_ratio calculation.

2003/09/04 02:23:31+01:00 davej
[CPUFREQ] Fix longhaul's mult,fsb -> MHz conversions.
this was wrong in several places by a factor of 100.
Introduce a calc_speed helper to make this harder to get wrong.
Additionally, clean up some printk's to print out useful values
like MinMult=3.5x instead of MinMult(x10)=35.

2003/09/04 01:48:51+01:00 davej
[CPUFREQ] Fix use of fsb before initialisation in longhaul.

2003/09/04 01:24:52+01:00 davej
[CPUFREQ] Cache FSB in longhaul driver.

2003/09/03 02:54:08+01:00 davej
[CPUFREQ] Inform user about broken powernow-k7 PST tables.

2003/09/03 02:27:43+01:00 davej
[CPUFREQ] fix speedstep-ich's .name to reflect its real name

2003/09/03 02:19:46+01:00 davej
[CPUFREQ] new cpufreq_driver->resume callback, needed (at least) for speedstep-smi.

2003/09/03 02:14:06+01:00 davej
[CPUFREQ] Fix incorrect entry in Kconfig

2003/09/03 02:03:50+01:00 davej
[CPUFREQ][ARM] allow for easier Kconfig usage on ARM, and more features for SA11x0 users.

2003/09/03 01:53:30+01:00 davej
[CPUFREQ][X86] Small fixes necessary to separate the governors from policies.

2003/09/03 01:47:33+01:00 davej
[CPUFREQ][SPARC64] Small fixes necessary to separate the governors from policies.

2003/09/03 01:35:15+01:00 davej
[CPUFREQ][SH] Small fixes necessary to separate the governors from policies.

2003/09/03 01:10:20+01:00 davej
[CPUFREQ][PPC] Small fixes necessary to separate the governors from policies.

2003/09/03 00:57:16+01:00 davej
[CPUFREQ] Add the "performance" and "powersave" governors as modules.
Also add a config option so that the default governor [which is loaded
and enabled at boot] is started. Note that if you select "performance"
here, the CPU will be set to the highest possible frequency even if
the CPU was booted at a lower frequency. If you select "userspace",
the frequency will stay the same.

2003/09/03 00:50:10+01:00 davej
[CPUFREQ] Completely separate governors from policies.
Governors are for CPUfreq drivers which have a ->target callback,
Policies are for CPUfreq drivers which have a ->setpolicy callback.
Also, the "hardwired" governors "powersave" and "performance" are removed.

2003/09/03 00:42:28+01:00 davej
[CPUFREQ] Fix various oddities in the userspace governer.

2003/09/03 00:14:53+01:00 davej
[CPUFREQ] Move drivers/cpufreq/userspace.c to drivers/cpufreq/cpufreq_userspace.c
Module names of 'userspace' aren't very informative.

2003/09/02 23:31:13+01:00 davej
[CPUFREQ] fix up two typos.
cpuferq hakcers cnat spel.

2003/09/02 23:07:47+01:00 davej
[CPUFREQ] Move kernel/cpufreq.c to drivers/cpufreq/cpufreq.c
Also remove $Id$ tag.
No other code change.

BKrev: 3f58e0d9s2BRLBr_lp6gX7SPGHQCoA